startvm.sh: Set number of cores and ram size based on host system
Currently, 'startvm.sh' set default number of cores and ram size of VM
as 16 and 32 GB, repectively.  It is somewhat too big VM for desktop or
laptop host system.  This commit modifies the script to check number of
cpus and size of ram physically equipped to the host and use its
fraction for VM.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>